moved existing simulation code to physics folder
