+++ src/software/backend/BUILD
@@ -67,7 +67,7 @@ cc_library(
     hdrs = ["simulator_backend.h"],
     deps = [
         ":backend",
-        "//software/backend/simulation:physics_simulator",
+        "//software/backend/simulation/physics:physics_simulator",
         "//software/util/logger",
         "//software/world",
     ],
