+++ src/software/backend/simulation/physics/physics_simulator.h
@@ -2,8 +2,8 @@
 
 #include <optional>
 
-#include "software/backend/simulation/physics_ball.h"
-#include "software/backend/simulation/physics_field.h"
+#include "software/backend/simulation/physics/physics_ball.h"
+#include "software/backend/simulation/physics/physics_field.h"
 #include "software/util/time/duration.h"
 #include "software/util/time/timestamp.h"
 #include "software/world/world.h"
