+++ src/software/backend/simulation/physics/physics_simulator_test.cpp
@@ -1,4 +1,4 @@
-#include "software/backend/simulation/physics_simulator.h"
+#include "software/backend/simulation/physics/physics_simulator.h"
 
 #include <gtest/gtest.h>
